move FluidIngredient to recipe package
