+++ src/main/java/de/dafuqs/spectrum/compat/REI/plugins/TitrationBarrelDisplay.java
@@ -2,7 +2,7 @@ package de.dafuqs.spectrum.compat.REI.plugins;
 
 import de.dafuqs.revelationary.api.advancements.*;
 import de.dafuqs.spectrum.compat.REI.*;
-import de.dafuqs.spectrum.helpers.FluidIngredient;
+import de.dafuqs.spectrum.recipe.*;
 import de.dafuqs.spectrum.recipe.titration_barrel.*;
 import me.shedaniel.rei.api.common.category.*;
 import me.shedaniel.rei.api.common.display.basic.*;
