+++ src/main/java/de/dafuqs/spectrum/compat/emi/recipes/TitrationBarrelEmiRecipeGated.java
@@ -1,7 +1,7 @@
 package de.dafuqs.spectrum.compat.emi.recipes;
 
 import de.dafuqs.spectrum.compat.emi.*;
-import de.dafuqs.spectrum.helpers.FluidIngredient;
+import de.dafuqs.spectrum.recipe.*;
 import de.dafuqs.spectrum.recipe.titration_barrel.*;
 import dev.emi.emi.api.stack.*;
 import dev.emi.emi.api.widget.TextWidget.*;
