+++ src/main/java/de/dafuqs/spectrum/recipe/GatedRecipeSerializer.java
@@ -1,13 +1,12 @@
 package de.dafuqs.spectrum.recipe;
 
 import com.google.gson.*;
-import de.dafuqs.spectrum.helpers.FluidIngredient;
 import net.minecraft.network.*;
 import net.minecraft.recipe.*;
 import net.minecraft.util.*;
 import org.jetbrains.annotations.*;
 
-import java.util.Objects;
+import java.util.*;
 
 public interface GatedRecipeSerializer<T extends Recipe<?>> extends RecipeSerializer<T> {
