+++ src/main/java/de/dafuqs/spectrum/recipe/titration_barrel/ITitrationBarrelRecipe.java
@@ -1,7 +1,6 @@
 package de.dafuqs.spectrum.recipe.titration_barrel;
 
 import de.dafuqs.spectrum.*;
-import de.dafuqs.spectrum.helpers.FluidIngredient;
 import de.dafuqs.spectrum.helpers.TimeHelper;
 import de.dafuqs.spectrum.recipe.*;
 import de.dafuqs.spectrum.registries.*;
