+++ src/main/java/de/dafuqs/spectrum/recipe/titration_barrel/TitrationBarrelRecipeSerializer.java
@@ -2,15 +2,11 @@ package de.dafuqs.spectrum.recipe.titration_barrel;
 
 import com.google.gson.*;
 import de.dafuqs.spectrum.*;
-import de.dafuqs.spectrum.helpers.FluidIngredient;
 import de.dafuqs.spectrum.recipe.*;
 import net.id.incubus_core.recipe.*;
-import net.id.incubus_core.util.RegistryHelper;
-import net.minecraft.fluid.*;
 import net.minecraft.item.*;
 import net.minecraft.network.*;
 import net.minecraft.registry.*;
-import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.*;
 
 import java.util.*;
